requirements.txt
streamlit
faiss-cpu
numpy
pandas
sentence-transformers
cohere
PyPDF2